action-allowedlist/pyproject.toml
[tool.poetry]
name = "action-allowedlist"
version = "0.1.0"
description = ""
authors = ["Christian <chrisgo@bu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
